DockPanel Simplification
