--- src/Perspex.Controls/DockPanel/Alignment.cs
@@ -1,10 +1,0 @@
-﻿namespace Perspex.Controls
-{
-    public enum Alignment
-    {
-        Stretch,
-        Start,
-        Middle,
-        End,
-    }
-}
