--- src/Perspex.Controls/DockPanel/Dock.cs
@@ -1,10 +1,0 @@
-namespace Perspex.Controls
-{
-    public enum Dock
-    {
-        Left = 0,
-        Bottom,
-        Right,
-        Top
-    }
-}
