--- src/Perspex.Controls/Margins.cs
@@ -1,8 +1,0 @@
-namespace Perspex.Controls
-{
-    public class Margins
-    {
-        public Segment HorizontalMargin { get; set; }
-        public Segment VerticalMargin { get; set; }
-    }
-}
