+++ src/Perspex.Controls/Perspex.Controls.csproj
@@ -41,17 +41,7 @@
     <Compile Include="..\Shared\SharedAssemblyInfo.cs">
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="DockPanel\Aligner.cs" />
-    <Compile Include="DockPanel\Alignment.cs" />
-    <Compile Include="DockPanel\Alignments.cs" />
-    <Compile Include="DockPanel\CoordinateMixin.cs" />
-    <Compile Include="DockPanel\Dock.cs" />
-    <Compile Include="DockPanel\Docker.cs" />
-    <Compile Include="DockPanel\DockingArranger.cs" />
-    <Compile Include="DockPanel\DockPanel.cs" />
-    <Compile Include="DockPanel\EnumerableMixin.cs" />
-    <Compile Include="DockPanel\LayoutSizes.cs" />
-    <Compile Include="DockPanel\RectMixin.cs" />
+    <Compile Include="DockPanel.cs" />
     <Compile Include="Generators\ITreeItemContainerGenerator.cs" />
     <Compile Include="Generators\ItemContainers.cs" />
     <Compile Include="IContentControl.cs" />
@@ -59,9 +49,6 @@
     <Compile Include="IPanel.cs" />
     <Compile Include="IReparentingHost.cs" />
     <Compile Include="ISetLogicalParent.cs" />
-    <Compile Include="DockPanel\LeftDocker.cs" />
-    <Compile Include="DockPanel\Segment.cs" />
-    <Compile Include="Margins.cs" />
     <Compile Include="MenuItemAccessKeyHandler.cs" />
     <Compile Include="Mixins\SelectableMixin.cs" />
     <Compile Include="Presenters\IContentPresenter.cs" />
@@ -79,7 +66,6 @@
     <Compile Include="Primitives\IScrollInfo.cs" />
     <Compile Include="Primitives\Popup.cs" />
     <Compile Include="Primitives\ScrollInfoAdapter.cs" />
-    <Compile Include="DockPanel\RightDocker.cs" />
     <Compile Include="Canvas.cs" />
     <Compile Include="Templates\ControlTemplate`2.cs" />
     <Compile Include="Templates\DataTemplate`1.cs" />
@@ -223,6 +209,7 @@
       <Name>Perspex.Styling</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
